module4/class3/pyproject.toml
[project]
name = "class3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.1.3",
    "langchain-openai>=1.1.3",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
]
